% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCorrections.R
\name{fetchCorrections}
\alias{fetchCorrections}
\title{Fetch Assignment Corrections}
\usage{
fetchCorrections(course_id, assignment_id)
}
\arguments{
\item{course_id}{Should match the subdirectory within GitHub (e.g., `"2025-sum-dsan5650"`)}

\item{assignment_id}{Should match the beginning of the corrections .md filename on GitHub}
}
\description{
Call this function to fetch and display the latest assignment corrections.
}
