% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagr-package.R
\docType{package}
\name{jagr}
\alias{jagr-package}
\alias{jagr}
\title{RStudio JAG Addins}
\description{
Use RStudio JAG Addins to submit to JAG from within RStudio.
}
\author{
\strong{Maintainer}: Jeff Jacobs \email{jj1088@georgetown.edu}

}
