% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCorrections.R
\name{fetchCorrections}
\alias{fetchCorrections}
\title{Fetch Assignment Corrections}
\usage{
fetchCorrections(assignment_id)
}
\arguments{
\item{assignment_id}{Should match the beginning of the corrections .md filename on GitHub}
}
\description{
Call this function to fetch and display the latest assignment corrections.
}
